+++ functions.php
@@ -606,6 +606,17 @@ function wrdsb_i_am_a_staff_site() {
   }
 }
 
+function wrdsb_i_am_a_school_site() {
+  $parsed_url = parse_url(site_url());
+  $host = explode('.', $parsed_url['host']);
+  $my_domains = array(
+    "schools",
+  );
+  if (in_array(($host[0]), $my_domains)) {
+    return TRUE;
+  }
+}
+
 function wrdsb_i_am_a_school() {
   $parsed_url = parse_url(site_url());
   $host = explode('.', $parsed_url['host']);